Move lenient-recompute functions out of mteb.api into _benchmark_metrics
_bucket_means/_recompute_lenient_means/_recompute_lenient_custom_groups
are pure functions over plain dicts, not tied to FastAPI/pydantic --
moved to mteb.benchmarks._benchmark_metrics (mteb.api.aggregators now
imports them) alongside the other aggregation helpers that already
live there.

Also extracted _bucket_task_result_scores, the TaskResult-based analog
of _bucket_means, and refactored _compute_task_types and
_compute_custom_group_means to share it instead of each re-implementing
the same bucket-by-key + null-tracking loop.

Tests moved from tests/test_api/ (now removed) into
test_benchmark_score.py alongside the other _benchmark_metrics.py
coverage. No behavior change -- 407 tests pass.
